requirements.txt
numpy
aocd
networkx
progressbar2

